WIP: fixed model relationships, added Beverage Model, change consumption type to Int(will be storing mls)

+++ Hydrate/Models/CoreData/Models/UserData.swift
@@ -6,7 +6,7 @@ final class UserData: NSManagedObject {
     fileprivate(set) var date: Date
 
     @NSManaged
-    fileprivate(set) var dailyGoal: Double
+    fileprivate(set) var dailyGoal: Int
 
     @NSManaged
     fileprivate(set) var weight: Double
